pyproject.toml
[tool.poetry]
authors = ["tqtensor <victortang212@gmail.com>"]
description = "Capture Users' Attention in 3s"
name = "three-seconds"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
ffmpy = "^0.3.1"
gdown = "^4.7.1"
pillow = "^10.2.0"
python = "^3.10"
transformers = {extras = ["torch"], version = "^4.36.2"}
youtube-dl = {git = "https://github.com/ytdl-org/youtube-dl.git", rev = "master"}

[tool.poetry.group.dev.dependencies]
dvc = {extras = ["gs"], version = "^2.58.2"}
jupyterlab = "==2.*"
nbconvert = "^7.11.0"
nbqa = {extras = ["toolchain"], version = "^1.7.0"}
poetry-plugin-export = "^1.6.0"
pre-commit = "^2.20.0"
toml-sort = "^0.23.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
